+++ Engine/EngineSource/3D Engine/Model/Meshes/AnimatedModel.h
@@ -110,7 +110,7 @@ namespace Animation
 
 		AnimationClip() : m_frameCount(0), m_framerate(24){};
 		AnimationClip(size_t frameCount, size_t jointCount)
-			: m_frameCount(frameCount), m_framerate(24)
+			: m_frameCount((uint16_t)frameCount), m_framerate(24)
 		{
 			m_skeletonPoses = std::make_unique<SkeletonPose[]>(frameCount);
 			for (int i = 0; i < jointCount; i++)
